Create HyBidAdView with AdSize
