+++ hybid.sdk/src/main/java/net/pubnative/lite/sdk/views/HyBidAdView.java
@@ -1,0 +1,237 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2020 PubNative GmbH
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+package net.pubnative.lite.sdk.views;
+
+import android.annotation.TargetApi;
+import android.content.Context;
+import android.text.TextUtils;
+import android.util.AttributeSet;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.RelativeLayout;
+
+import net.pubnative.lite.sdk.api.RequestManager;
+import net.pubnative.lite.sdk.models.Ad;
+import net.pubnative.lite.sdk.models.AdSize;
+import net.pubnative.lite.sdk.models.IntegrationType;
+import net.pubnative.lite.sdk.presenter.AdPresenter;
+import net.pubnative.lite.sdk.utils.Logger;
+
+public abstract class HyBidAdView extends RelativeLayout implements RequestManager.RequestListener, AdPresenter.Listener {
+
+    public interface Listener {
+        void onAdLoaded();
+
+        void onAdLoadFailed(Throwable error);
+
+        void onAdImpression();
+
+        void onAdClick();
+    }
+
+    private RequestManager mRequestManager;
+    protected PNAdView.Listener mListener;
+    private AdPresenter mPresenter;
+    protected Ad mAd;
+    private AdSize mAdSize;
+
+    public HyBidAdView(Context context) {
+        super(context);
+        init(getRequestManager());
+    }
+
+    public HyBidAdView(Context context, AttributeSet attrs) {
+        super(context, attrs);
+        init(getRequestManager());
+    }
+
+    public HyBidAdView(Context context, AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
+        init(getRequestManager());
+    }
+
+    @TargetApi(21)
+    public HyBidAdView(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+        super(context, attrs, defStyleAttr, defStyleRes);
+        init(getRequestManager());
+    }
+
+    private void init(RequestManager requestManager) {
+        mRequestManager = requestManager;
+        mRequestManager.setIntegrationType(IntegrationType.STANDALONE);
+        mAdSize = AdSize.SIZE_320x50;
+    }
+
+    public void setAdSize(AdSize adSize) {
+        this.mAdSize = adSize;
+    }
+
+    public void load(String zoneId, PNAdView.Listener listener) {
+        load(zoneId, mAdSize, listener);
+    }
+
+    public void load(String zoneId, AdSize adSize, PNAdView.Listener listener) {
+        cleanup();
+        mAdSize = adSize;
+        mListener = listener;
+        if (TextUtils.isEmpty(zoneId)) {
+            invokeOnLoadFailed(new Exception("Invalid zone id provided"));
+        } else {
+            mRequestManager.setZoneId(zoneId);
+            mRequestManager.setRequestListener(this);
+            mRequestManager.requestAd();
+        }
+    }
+
+    public void destroy() {
+        cleanup();
+        if (mRequestManager != null) {
+            mRequestManager.destroy();
+            mRequestManager = null;
+        }
+    }
+
+    protected void cleanup() {
+        stopTracking();
+        removeAllViews();
+        mAd = null;
+
+        if (mPresenter != null) {
+            mPresenter.destroy();
+            mPresenter = null;
+        }
+    }
+
+    public String getImpressionId() {
+        return mAd != null ? mAd.getImpressionId() : null;
+    }
+
+    public String getCreativeId() {
+        return mAd != null ? mAd.getCreativeId() : null;
+    }
+
+    protected abstract String getLogTag();
+
+    abstract RequestManager getRequestManager();
+
+    protected abstract AdPresenter createPresenter();
+
+    protected void renderAd() {
+        mPresenter = createPresenter();
+        if (mPresenter != null) {
+            mPresenter.load();
+        } else {
+            invokeOnLoadFailed(new Exception("The server has returned an unsupported ad asset"));
+        }
+    }
+
+    protected void startTracking() {
+        if (mPresenter != null) {
+            mPresenter.startTracking();
+        }
+    }
+
+    protected void stopTracking() {
+        if (mPresenter != null) {
+            mPresenter.stopTracking();
+        }
+    }
+
+    protected void invokeOnLoadFinished() {
+        if (mListener != null) {
+            mListener.onAdLoaded();
+        }
+    }
+
+    protected void invokeOnLoadFailed(Exception exception) {
+        Logger.e(getLogTag(), exception.getMessage());
+        if (mListener != null) {
+            mListener.onAdLoadFailed(exception);
+        }
+    }
+
+    protected void invokeOnClick() {
+        if (mListener != null) {
+            mListener.onAdClick();
+        }
+    }
+
+    protected void invokeOnImpression() {
+        if (mListener != null) {
+            mListener.onAdImpression();
+        }
+    }
+
+    protected void setupAdView(View view) {
+        RelativeLayout.LayoutParams adLayoutParams = new RelativeLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT);
+        adLayoutParams.addRule(RelativeLayout.CENTER_IN_PARENT, RelativeLayout.TRUE);
+
+        addView(view, adLayoutParams);
+
+        invokeOnLoadFinished();
+        startTracking();
+        invokeOnImpression();
+    }
+
+    public void setMediation(boolean isMediation) {
+        if (mRequestManager != null) {
+            mRequestManager.setIntegrationType(isMediation ? IntegrationType.MEDIATION : IntegrationType.STANDALONE);
+        }
+    }
+
+    //----------------------------- AdPresenter Callbacks --------------------------------------
+    @Override
+    public void onRequestSuccess(Ad ad) {
+        if (ad == null) {
+            invokeOnLoadFailed(new Exception("Server returned null ad"));
+        } else {
+            mAd = ad;
+            renderAd();
+        }
+    }
+
+    @Override
+    public void onRequestFail(Throwable throwable) {
+        invokeOnLoadFailed(new Exception(throwable));
+    }
+
+    //----------------------------- AdPresenter Callbacks --------------------------------------
+    @Override
+    public void onAdLoaded(AdPresenter adPresenter, View banner) {
+        if (banner == null) {
+            invokeOnLoadFailed(new Exception("An error has occurred while rendering the ad"));
+        } else {
+            setupAdView(banner);
+        }
+    }
+
+    @Override
+    public void onAdError(AdPresenter adPresenter) {
+        invokeOnLoadFailed(new Exception("An error has occurred while rendering the ad"));
+    }
+
+    @Override
+    public void onAdClicked(AdPresenter adPresenter) {
+        invokeOnClick();
+    }
+}
