+++ hybid.sdk/src/main/java/net/pubnative/lite/sdk/views/PNAdView.java
@@ -24,202 +24,25 @@ package net.pubnative.lite.sdk.views;
 
 import android.annotation.TargetApi;
 import android.content.Context;
-import android.text.TextUtils;
 import android.util.AttributeSet;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.RelativeLayout;
 
-import net.pubnative.lite.sdk.api.RequestManager;
-import net.pubnative.lite.sdk.models.Ad;
-import net.pubnative.lite.sdk.models.IntegrationType;
-import net.pubnative.lite.sdk.presenter.AdPresenter;
-import net.pubnative.lite.sdk.utils.Logger;
-
-public abstract class PNAdView extends RelativeLayout implements RequestManager.RequestListener, AdPresenter.Listener {
-
-    public interface Listener {
-        void onAdLoaded();
-
-        void onAdLoadFailed(Throwable error);
-
-        void onAdImpression();
-
-        void onAdClick();
-    }
-
-    private RequestManager mRequestManager;
-    protected Listener mListener;
-    private AdPresenter mPresenter;
-    protected Ad mAd;
+public abstract class PNAdView extends HyBidAdView {
+    public interface Listener extends HyBidAdView.Listener { }
 
     public PNAdView(Context context) {
         super(context);
-        init(getRequestManager());
     }
 
     public PNAdView(Context context, AttributeSet attrs) {
         super(context, attrs);
-        init(getRequestManager());
     }
 
     public PNAdView(Context context, AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
-        init(getRequestManager());
     }
 
     @TargetApi(21)
     public PNAdView(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
         super(context, attrs, defStyleAttr, defStyleRes);
-        init(getRequestManager());
-    }
-
-    private void init(RequestManager requestManager) {
-        mRequestManager = requestManager;
-        mRequestManager.setIntegrationType(IntegrationType.STANDALONE);
-    }
-
-    public void load(String zoneId, Listener listener) {
-        cleanup();
-        mListener = listener;
-        if (TextUtils.isEmpty(zoneId)) {
-            invokeOnLoadFailed(new Exception("Invalid zone id provided"));
-        } else {
-            mRequestManager.setZoneId(zoneId);
-            mRequestManager.setRequestListener(this);
-            mRequestManager.requestAd();
-        }
-    }
-
-    public void destroy() {
-        cleanup();
-        if (mRequestManager != null) {
-            mRequestManager.destroy();
-            mRequestManager = null;
-        }
-    }
-
-    protected void cleanup() {
-        stopTracking();
-        removeAllViews();
-        mAd = null;
-
-        if (mPresenter != null) {
-            mPresenter.destroy();
-            mPresenter = null;
-        }
-    }
-
-    public String getImpressionId() {
-        return mAd != null ? mAd.getImpressionId() : null;
-    }
-
-    public String getCreativeId() {
-        return mAd != null ? mAd.getCreativeId() : null;
-    }
-
-    protected abstract String getLogTag();
-
-    abstract RequestManager getRequestManager();
-
-    protected abstract AdPresenter createPresenter();
-
-    protected void renderAd() {
-        mPresenter = createPresenter();
-        if (mPresenter != null) {
-            mPresenter.load();
-        } else {
-            invokeOnLoadFailed(new Exception("The server has returned an unsupported ad asset"));
-        }
-    }
-
-    protected void startTracking() {
-        if (mPresenter != null) {
-            mPresenter.startTracking();
-        }
-    }
-
-    protected void stopTracking() {
-        if (mPresenter != null) {
-            mPresenter.stopTracking();
-        }
-    }
-
-    protected void invokeOnLoadFinished() {
-        if (mListener != null) {
-            mListener.onAdLoaded();
-        }
-    }
-
-    protected void invokeOnLoadFailed(Exception exception) {
-        Logger.e(getLogTag(), exception.getMessage());
-        if (mListener != null) {
-            mListener.onAdLoadFailed(exception);
-        }
-    }
-
-    protected void invokeOnClick() {
-        if (mListener != null) {
-            mListener.onAdClick();
-        }
-    }
-
-    protected void invokeOnImpression() {
-        if (mListener != null) {
-            mListener.onAdImpression();
-        }
-    }
-
-    protected void setupAdView(View view) {
-        LayoutParams adLayoutParams = new LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT);
-        adLayoutParams.addRule(RelativeLayout.CENTER_IN_PARENT, RelativeLayout.TRUE);
-
-        addView(view, adLayoutParams);
-
-        invokeOnLoadFinished();
-        startTracking();
-        invokeOnImpression();
-    }
-
-    public void setMediation(boolean isMediation) {
-        if (mRequestManager != null) {
-            mRequestManager.setIntegrationType(isMediation ? IntegrationType.MEDIATION : IntegrationType.STANDALONE);
-        }
-    }
-
-    //----------------------------- AdPresenter Callbacks --------------------------------------
-    @Override
-    public void onRequestSuccess(Ad ad) {
-        if (ad == null) {
-            invokeOnLoadFailed(new Exception("Server returned null ad"));
-        } else {
-            mAd = ad;
-            renderAd();
-        }
-    }
-
-    @Override
-    public void onRequestFail(Throwable throwable) {
-        invokeOnLoadFailed(new Exception(throwable));
-    }
-
-    //----------------------------- AdPresenter Callbacks --------------------------------------
-    @Override
-    public void onAdLoaded(AdPresenter adPresenter, View banner) {
-        if (banner == null) {
-            invokeOnLoadFailed(new Exception("An error has occurred while rendering the ad"));
-        } else {
-            setupAdView(banner);
-        }
-    }
-
-    @Override
-    public void onAdError(AdPresenter adPresenter) {
-        invokeOnLoadFailed(new Exception("An error has occurred while rendering the ad"));
-    }
-
-    @Override
-    public void onAdClicked(AdPresenter adPresenter) {
-        invokeOnClick();
     }
 }
